refactor: optionBox
